+++ src/contributions/_index.ts
@@ -3,10 +3,12 @@ import registerCommands from "./commands/_index";
 import registerCodeLensProviders from "./codelens/_index";
 import registerEditorProviders from "./editors/_index";
 import registerLanguageProviders from "./language/_index";
+import registerStatusBarItems from "./status-bars/_index";
 
 export default function registerContributions(context: vscode.ExtensionContext) {
 	registerCommands();
 	registerCodeLensProviders();
 	registerLanguageProviders(context);
 	registerEditorProviders(context);
+	registerStatusBarItems(context);
 }