+++ src/contributions/status-bars/_index.ts
@@ -1,0 +1,6 @@
+import * as vscode from "vscode";
+import registerRunBuildStatusBarItem from "./run-build-item";
+
+export default function registerStatusBarItems(context: vscode.ExtensionContext) {
+  registerRunBuildStatusBarItem(context);
+}